feat(knowledge): B-KNOWLEDGE-TAXONOMY-01 P5 — nested circle-pack knowledge page [flag-off]
The knowledge-bubbles direction behind KNOWLEDGE_MAP_PAGE (default off):

- Flag gate: page.tsx is now a thin server wrapper. OFF renders
  KnowledgeFlatClient — the previous client page moved VERBATIM (diff
  proves exactly one line changed: the export declaration); the map
  loads via dynamic import so its code never ships on the off path.
- knowledge-tree.ts: owned leaves (real mastery via
  getKnowledgePageData) assembled into the authored graph. Ghosts =
  unheld roster siblings with a layout-only footprint — sumRealPoints
  excludes them (§5.1: they never inflate a parent). One home-parent
  containment (§E, first substantive edge). Leaf mastery leaf-exact;
  parent mastery via parentProgress (threshold + >=2 corners). Owned
  leaves the graph doesn't know land under the root, so the map is
  complete while the graph is sparse.
- KnowledgeBubbleMap.tsx: d3-hierarchy pack (the only new dep —
  tree-shakeable, no full d3); zoom traversal + clickable breadcrumb;
  peaks strip (top-5, tap → focus + gold flash); ghost tap → the real
  territory-adopt endpoint (domain-frequency 'sometimes'), optimistic
  with revert + quiet inline notice; list-view toggle; mastery = gold
  fill PLUS the 'Mastery' label (never color alone); field hues on the
  --cat-* scale, always paired with the name; keyboard-focusable
  bubbles, prefers-reduced-motion disables zoom/flash animation.

7 tree tests (ghosts-never-inflate, home-parent single containment,
leaf-exact mastery, corner gate, ungrouped leaves); 49 green in
src/server/knowledge. Ratchets all pass; typecheck + lint clean.
Follow-ups noted: share/tidy chrome on the map page, friend page,
collection-parent rendering, grow-rim.
