+++ package.json
@@ -43,6 +43,7 @@
     "bcryptjs": "latest",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
+    "d3-hierarchy": "^3.1.2",
     "dotenv": "^17.4.2",
     "drizzle-orm": "^0.45.2",
     "html2canvas": "^1.4.1",
@@ -64,6 +65,7 @@
   "devDependencies": {
     "@tailwindcss/postcss": "latest",
     "@types/bcryptjs": "latest",
+    "@types/d3-hierarchy": "^3.1.7",
     "@types/node": "latest",
     "@types/pg": "^8.20.0",
     "@types/react": "latest",
